pyproject.toml
[project]
name = "deckeditor"

[tool.isort]
profile = "black"
lines_after_imports = 2
honor_noqa = true

[tool.black]
line-length = 119

[tool.poetry]
name = "deckeditor"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "~3.9"
PyQt5 = "5.15.10"
PyQt5-Qt5 = "5.15.2"
requests = "^2.31.0"
promise = "~2.3"
notify-py = "^0.3.42"
frozendict = "~1.2"
bidict = "^0.23.0"
simplejson = "~3.19"
sqlalchemy = "~2.0"
sortedcontainers = "^2.4.0"
gevent = "~24.2"
flask = "^3.0.2"
sqlalchemy-jsonfield = "^1.0.2"
lru-dict = "^1.3.0"
Pillow = "~9.4"
appdirs = "~1.4"
pysqlite3 = "~0.5"
cubeclient = { git = "https://github.com/guldfisk/cubeclient.git" }
orp = { git = "https://github.com/guldfisk/orp.git" }
mtgorp = { git = "https://github.com/guldfisk/mtgorp.git" }
mtgqt = { git = "https://github.com/guldfisk/mtgqt.git" }
mtgimg = { git = "https://github.com/guldfisk/mtgimg.git" }
magiccube = { git = "https://github.com/guldfisk/magiccube.git" }
mtgdraft = { git = "https://github.com/guldfisk/mtgdraft.git" }
hardcandy = { git = "https://github.com/guldfisk/hardcandy.git" }
yeetlong = { git = "https://github.com/guldfisk/yeetlong.git" }
lobbyclient = { git = "https://github.com/guldfisk/lobbyclient.git" }


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"
PyQt5-stubs = "~5.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
